split main.js to modules
